small fixes on fx
echo for some reason still happens until you move the slider up a little
then back to 0. I think because even though we have the values
initialized under riff.cs , the audio filters never gets updated until
we touch the sliders so until we move those sliders they get the default
numbers that unity gives.  so we can either figure out how to change the
default numbers on some or just turn off the filters from the start and
have the players turn it on themselves